Implement Redis store with sorted sets and execution engine integration. Refactored hashmap, skip list, and buffer implementations

+++ server/include/parser/resp_parser.h
@@ -21,7 +21,7 @@ typedef struct BulkString {
 
 typedef struct RedisCommand {
     struct BulkString *args;
-    size_t arg_count;
+    int arg_count;
 } RedisCommand;
 
 /* Function Prototypes */
